moves grid/particle dataset plot plugin to new file
