+++ nata/plugins/plot/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 from .grid import *
 from .particle import *
-
+from .dataset import *
 # from .collection import *
